+++ packages/ncform-e2e/ncform/cypress/integration/examples/layoutWidgets.spec.js
@@ -62,6 +62,16 @@ context('Layout widgets', () => {
     })
   })
 
+  it('array: some rows can not be deleted', () => {
+    let id = md5('array: some rows can not be deleted');
+    cy.get(`[data-cy=${id}]`).within(() => {
+      cy.get('legend').contains('user').as('userLegend');
+
+      cy.get('@userLegend').next().find('button.btn-danger').should('not.exist');
+      cy.get('@userLegend').next().next().find('button.btn-danger').should('exist');
+    })
+  })
+
   it('array: collapsed / itemCollapse true and txts change', () => {
     let id = md5('array: collapsed / itemCollapse true and txts change');
     cy.get(`[data-cy=${id}]`).within(() => {
@@ -104,6 +114,37 @@ context('Layout widgets', () => {
 
       cy.get('@userLegend').next().find('button').contains('Add Item').should('exist');
       cy.get('@userLegend').next().find('button').contains('Remove All').should('exist');
+    })
+  })
+
+  it('table: some rows can be deleted', () => {
+    let id = md5('table: some rows can be deleted');
+    cy.get(`[data-cy=${id}]`).within(() => {
+      cy.get('legend').contains('user').as('userLegend');
+
+      cy.get('@userLegend').next().find('button.btn-danger').should('have.length', 1);
+    })
+  })
+
+  it('table: set colums width', () => {
+    let id = md5('table: set colums width');
+    cy.get(`[data-cy=${id}]`).within(() => {
+      cy.get('thead th:last-child').should('have.prop', 'offsetWidth', 200);
+    })
+  })
+
+  it('array: showOneIfEmpty is true - No extra blank row when there has initial value', () => {
+    let id = md5('array: showOneIfEmpty is true - No extra blank row when there has initial value');
+    cy.get(`[data-cy=${id}]`).within(() => {
+      cy.get('input').should('have.length', 1);
+    })
+  })
+
+  it('array: with number items', () => {
+    let id = md5('array: with number items');
+    cy.get(`[data-cy=${id}]`).within(() => {
+      cy.get('legend').contains('user').as('userLegend');
+      cy.get('@userLegend').parent().find('button').contains('Add').click();
     })
   })