+++ packages/ncform-theme-elementui/conf/webpack.config.prd.js
@@ -1,23 +1,37 @@
 const path = require("path");
-const webpack = require("webpack");
+const merge = require("webpack-merge");
 
 const config = require("./config");
+const devConfig = require("./webpack.config.dev");
 
-module.exports = {
+const commonConfig = {
+  mode: "production",
 
-  mode: 'production',
-
-  devtool: false,
+  devtool: "source-map",
 
   entry: {
-    ncformStdComps: path.join(config.src, 'components', 'index.js'),
+    ncformStdComps: path.join(config.src, "components", "index.js")
   },
 
   output: {
-    filename: "[name].min.js",
+    filename: "[name].min.js"
   },
 
   optimization: {
     minimize: true
   }
 };
+
+module.exports = module.exports = [
+  // 非压缩版本
+  merge(devConfig[1], commonConfig, {
+    output: {
+      filename: "[name].js"
+    },
+    optimization: {
+      minimize: false
+    }
+  }),
+  // 压缩版本
+  merge(devConfig[1], commonConfig)
+];