+++ packages/ncform-theme-elementui/src/components/layout-comps/object.vue
@@ -10,12 +10,12 @@
 
       <div v-for="(fieldSchema, field) in schema.properties"
           :key="field"
-          :class="['el-col-' + (fieldSchema.ui.columns * 2 || 24)]"
+          :class="['el-col-' + ((_analyzeVal(fieldSchema.ui.columns) || 12) * 2 || 24)]"
           :style="{display: _analyzeVal(fieldSchema.ui.hidden) ? 'none' : ''}"
           class="el-col el-form-item">
 
         <template>
-            <label v-if="!fieldSchema.ui.noLabelSpace" :style="{'visibility': fieldSchema.ui.showLabel ? 'visible' : 'hidden'}" class="el-form-item__label">
+            <label v-if="!fieldSchema.ui.noLabelSpace" :style="{'visibility': fieldSchema.ui.showLabel ? 'visible' : 'hidden'}" class="el-form-item__label" :title="_analyzeVal(fieldSchema.ui.label)">
               <!-- 必填标识 -->
               <i v-if="_analyzeVal(fieldSchema.rules.required) === true || (typeof fieldSchema.rules.required === 'object' && _analyzeVal(fieldSchema.rules.required.value) === true)" class="text-danger">*</i>
               {{_analyzeVal(fieldSchema.ui.label)}}
@@ -43,11 +43,11 @@
     <div v-if="mergeConfig.layout === 'h'" v-show="!collapsed" class="el-row h-layout" style="width: 100%">
       <div v-for="(fieldSchema, field) in schema.properties"
           :key="field"
-          :class="['el-col-' + (fieldSchema.ui.columns * 2 || 24)]"
+          :class="['el-col-' + ((_analyzeVal(fieldSchema.ui.columns) || 12) * 2 || 24)]"
           :style="{display: _analyzeVal(fieldSchema.ui.hidden) ? 'none' : ''}"
           class="el-col el-form-item">
         <template>
-          <label v-if="!fieldSchema.ui.noLabelSpace" :style="{'visibility': fieldSchema.ui.showLabel ? 'visible' : 'hidden', width: mergeConfig.labelWidth}"  class="el-form-item__label">
+          <label v-if="!fieldSchema.ui.noLabelSpace" :style="{'visibility': fieldSchema.ui.showLabel ? 'visible' : 'hidden', width: mergeConfig.labelWidth}"  class="el-form-item__label" :title="_analyzeVal(fieldSchema.ui.label)">
             <!-- 必填标识 -->
             <i v-if="_analyzeVal(fieldSchema.rules.required) === true || (typeof fieldSchema.rules.required === 'object' && _analyzeVal(fieldSchema.rules.required.value) === true)" class="text-danger">*</i>
             {{_analyzeVal(fieldSchema.ui.label)}}
@@ -136,8 +136,23 @@
       .el-form-item {
         margin-bottom: 22px;
         .el-form-item__content {
-          line-height: 1.5;
+          line-height: unset;
         }
+      }
+    }
+
+    // 解决对象水平布局的组件显示问题
+    .h-layout {
+      .__ncform-control {
+        line-height: 40px;
+        &.__array-form-item, &.__array-table-form-item, &.__array-tabs-form-item {
+          line-height: unset;
+        }
+      }
+    }
+    .v-layout {
+      .__ncform-control {
+        line-height: unset;
       }
     }
   }