+++ packages/ncform/conf/webpack.config.prd.js
@@ -1,17 +1,30 @@
-const path = require("path");
-const webpack = require("webpack");
+const merge = require("webpack-merge");
+const devConfig = require("./webpack.config.dev");
 
-const config = require("./config");
-
-module.exports = {
-
-  mode: 'production',
+const commonConfig = {
+  mode: "production",
 
   output: {
-    filename: "[name].min.js",
+    filename: "[name].min.js"
   },
+
+  devtool: "source-map",
 
   optimization: {
     minimize: true
   }
 };
+
+module.exports = [
+  // 非压缩版本
+  merge(devConfig, commonConfig, {
+    output: {
+      filename: "[name].js"
+    },
+    optimization: {
+      minimize: false
+    }
+  }),
+  // 压缩版本
+  merge(devConfig, commonConfig)
+];